GreaterWrong: Fix iteration to not get stuck on the last item.
Unclear what circumstances cause this to occur,
but sometimes AF returns the item you moved the after date to,
which causes iteration to loop without the date advancing until the job is killed.

This adds a check for the exact case which simply returns,
as well as a more general check for failure to progress which throws.